pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dist-img-gen"
version = "0.1.0"
description = ""
authors = [
    { name = "Shih-Ying Yeh(KohakuBlueLeaf)", email = "apolloyeh0123@gmail.com" },
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "httpx",
    "peewee",
    "pillow"
]
urls = { "Homepage" = "https://kblueleaf.net" }

[project.optional-dependencies]
dev = []

[tool.setuptools]
package-dir = {"" = "src"}
